backend/bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = """* bot is a discord bot to make discord app bertter to:
    * advanced search features for who want make some records or fix the broken preview from some outside URLs.
    * Fix Links Previews."""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.5.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
]
